restructure into new directories
