+++ corehq/apps/reports_core/templates/reports_core/base_template_new.html
