+++ corehq/apps/reports_core/templates/reports_core/filter_new.html
